requirements.txt
hcl2
pyyaml
